requirements.txt
pandas
requests
psycopg2-binary
sqlalchemy
streamlit  # Corregido desde "stremalit"
folium
streamlit-folium
